02_APP/fastapi-app/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
pillow
pillow-heif
